renaming ODEParameters -> AbstractODEParams, moving odes.py out of model_configuration module and into base dynode, moving ODEParamsSEIP specification into the model_odes module.
